Fixed two issues:
#1 For the diff analysis we verify the file is part of the build spec
list, this was case sensitive. On windows we were seeing an issue were
the git diff returned a path with a different case to the file system.
As windows is case in-sensitive anyway, I have changed our comparison
to be case in-sensitive for paths in a windows format.

#2 Also with the diff analysis, the Jenkins toWindowsCommand() call was
incorrectly handling the git diff command to put quotes around the pipe
to a file. e.g. git diff <rev> ">" diff_list.txt. I have changed this to
just use the argument list builder and add the prefix/suffix as we do in
the same style for Linux.